Fixed the and tested the libraries

+++ lib/Navigator/Navigator.cpp
@@ -12,7 +12,7 @@ Navigator::Navigator(Motor& LMOTOR, Motor&  RMOTOR)
 * Negative PWM values will drive the robot in the opposite direction.
 * @param PWM PWM that will be sent to the PWM channel.
 */
-void Navigator::moveAhead(const int16_t PWM) {
+void Navigator::moveAhead(const int PWM) {
 
 	if (PWM < 0) {Navigator::moveBackwards(-PWM);}
 	else {
@@ -27,7 +27,7 @@ void Navigator::moveAhead(const int16_t PWM) {
 * Negative PWM values will drive the robot in the opposite direction.
 * @param PWM PWM that will be sent to the PWM channel.
 */
-void Navigator::moveBackwards(const int16_t PWM) {
+void Navigator::moveBackwards(const int PWM) {
 
 	if (PWM < 0) {Navigator::moveAhead(-PWM);}
 	else {
@@ -42,7 +42,7 @@ void Navigator::halt() {
 	Navigator::changeSpeed(0);
 }
 
-void Navigator::turnLeft(const int16_t PWM) {
+void Navigator::turnLeft(const int PWM) {
 
 	if (PWM < 0) {Navigator::turnRight(-PWM);}
 	else {
@@ -52,7 +52,7 @@ void Navigator::turnLeft(const int16_t PWM) {
 	}
 }
 
-void Navigator::turnRight(const int16_t PWM) {
+void Navigator::turnRight(const int PWM) {
 
 	if (PWM < 0) {Navigator::turnRight(-PWM);}
 	else {
@@ -62,7 +62,7 @@ void Navigator::turnRight(const int16_t PWM) {
 	}
 }
 
-void Navigator::changeSpeed(const int16_t PWM) {
+void Navigator::changeSpeed(const int PWM) {
 
 	lMotor.setSpeed(PWM);
 	rMotor.setSpeed(PWM);
